fix 订单 触发 订单配送待测试（履约 存在问题, AT 事务 本地事务 回滚不了问题待研究）

+++ taki-mall-order/src/main/java/com/taki/order/domain/dto/OrderListQueryDTO.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import org.apache.commons.lang3.tuple.Pair;
 import org.springframework.util.CollectionUtils;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.Set;
 
@@ -77,7 +78,7 @@ public class OrderListQueryDTO {
     /**
      * 支付金额-查询区间
      */
-    private Pair<Integer,Integer> payAmountInterval;
+    private Pair<BigDecimal,BigDecimal> payAmountInterval;
 
 
     /**
